- set default values in local copy of model to match Hayloft - fix to "reheatType" config loading that was lost during a previous merge

+++ app/src/main/java/a75f/io/renatus/profiles/vav/VavProfileConfigFragment.kt
@@ -165,8 +165,8 @@ class VavProfileConfigFragment : BaseDialogFragment() {
                         list = viewModel.reheatTypesList,
                         previewWidth = 160,
                         expandedWidth = 160,
-                        onSelected = { selectedIndex -> viewModel.viewState.reheatType = selectedIndex},
-                        defaultSelection = viewModel.viewState.reheatType
+                        onSelected = { selectedIndex -> viewModel.viewState.reheatType = selectedIndex.toDouble()},
+                        defaultSelection = viewModel.viewState.reheatType.toInt()
                     )
                     DropDownWithLabel(
                         label = "Zone Priority",
